+++ plugins/strangemood.cpp
@@ -574,7 +574,7 @@ command_result df_strangemood (color_ostream &out, vector <string> & parameters)
     // Randomly select a unit to enter a mood
     if (!unit)
     {
-        vector<int32_t> tickets;
+        vector<size_t> tickets;
         for (size_t i = 0; i < moodable_units.size(); i++)
         {
             df::unit *cur = moodable_units[i];
@@ -1119,7 +1119,7 @@ command_result df_strangemood (color_ostream &out, vector <string> & parameters)
          (job->job_type == job_type::StrangeMoodFell)
        ))
     {
-        int extra_items = std::min(rng.df_trandom((ui->tasks.num_artifacts * 20 + moodable_units.size()) / 20 + 1), 7);
+        size_t extra_items = std::min(rng.df_trandom((ui->tasks.num_artifacts * 20 + moodable_units.size()) / 20 + 1), 7);
         df::item_type avoid_type = item_type::NONE;
         int avoid_glass = 0;
         switch (skill)
